dark_spot_counter.py
Performs HyperSpy find_peaks_log() and then applies circle markers to the peak positions on the Signal2D. Secondly, interactive marker placement and removal is enabled by mouse clicking inside the plot. Either of the functionalities can be used independently. Using the signal scale and the number of markers, the density of dark spots is calculated and applied to the signal metadata.